Add basic tests of SpiceMix functionality.
These tests are based on the current implementation of the algorithm
and are subject to change. The input data for these tests must eventually
either be uploaded to GitHub (probably not a good idea if they are too large)
or placed in some other downloadable location (e.g. AWS bucket).